convert to psake

+++ psakefile.ps1
@@ -1,12 +1,26 @@
 Task Default -Depends Link
 
 Task Link {
+    # -----[ bin ]-------------------------------
+    Get-ChildItem -File -Path 'bin' | %{
+        $link = Join-Path (Resolve-Path '~') $_.Name
+        $target = $_.FullName
+
+        if (Test-Path $link) {
+            Remove-Item -Force -Path $link | Out-Null
+        }
+
+        cmd /c mklink $link $target
+    }
+
     # -----[ dotfiles ]--------------------------
     Get-ChildItem -File -Path 'src' | %{
         $link = Join-Path (Resolve-Path '~') $_.Name
         $target = $_.FullName
 
-        Remove-Item -Force -Path $link | Out-Null
+        if (Test-Path $link) {
+            Remove-Item -Force -Path $link | Out-Null
+        }
 
         cmd /c mklink $link $target
     }
@@ -20,7 +34,9 @@ Task Link {
         $link = Join-Path $gpgdir $_.Name
         $target = $_.FullName
 
-        Remove-Item -Force -Path $link | Out-Null
+        if (Test-Path $link) {
+            Remove-Item -Force -Path $link | Out-Null
+        }
 
         cmd /c mklink $link $target
     }
@@ -34,7 +50,9 @@ Task Link {
         $link = Join-Path $svndir $_.Name
         $target = $_.FullName
 
-        Remove-Item -Force -Path $link | Out-Null
+        if (Test-Path $link) {
+            Remove-Item -Force -Path $link | Out-Null
+        }
 
         cmd /c mklink $link $target
     }
@@ -48,7 +66,9 @@ Task Link {
         $link = Join-Path $subldir $_.Name
         $target = $_.FullName
 
-        Remove-Item -Force -Path $link | Out-Null
+        if (Test-Path $link) {
+            Remove-Item -Force -Path $link | Out-Null
+        }
 
         cmd /c mklink $link $target
     }
